refactor(test): move divider and switch inline tests to integration tests
