+++ src/widget/divider.rs
@@ -253,107 +253,4 @@ pub fn vdivider() -> Divider {
     Divider::vertical()
 }
 
-// Most tests moved to tests/widget_tests.rs
-// Tests below access private fields and must stay inline
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_divider_new() {
-        let d = Divider::new();
-        assert_eq!(d.orientation, Orientation::Horizontal);
-        assert_eq!(d.style, DividerStyle::Solid);
-    }
-
-    #[test]
-    fn test_divider_vertical() {
-        let d = Divider::vertical();
-        assert_eq!(d.orientation, Orientation::Vertical);
-    }
-
-    #[test]
-    fn test_divider_builder() {
-        let d = divider().dashed().color(Color::RED).label("Test").margin(2);
-
-        assert_eq!(d.style, DividerStyle::Dashed);
-        assert_eq!(d.label, Some("Test".to_string()));
-        assert_eq!(d.margin, 2);
-    }
-
-    #[test]
-    fn test_line_chars() {
-        let d = divider();
-        assert_eq!(d.line_char(), '─');
-
-        let d = divider().dashed();
-        assert_eq!(d.line_char(), '╌');
-
-        let d = divider().double();
-        assert_eq!(d.line_char(), '═');
-
-        let d = vdivider();
-        assert_eq!(d.line_char(), '│');
-
-        let d = vdivider().thick();
-        assert_eq!(d.line_char(), '┃');
-    }
-
-    #[test]
-    fn test_helper_functions() {
-        let h = divider();
-        assert_eq!(h.orientation, Orientation::Horizontal);
-
-        let v = vdivider();
-        assert_eq!(v.orientation, Orientation::Vertical);
-    }
-
-    #[test]
-    fn test_divider_render_uses_helpers() {
-        // Test that divider rendering works correctly after migration to helpers
-        let mut buffer = Buffer::new(30, 1);
-        let area = Rect::new(0, 0, 30, 1);
-        let mut ctx = RenderContext::new(&mut buffer, area);
-
-        let d = divider().label("Test");
-        d.render(&mut ctx);
-
-        // Verify label is rendered
-        let text: String = (0..30)
-            .filter_map(|x| buffer.get(x, 0).map(|c| c.symbol))
-            .collect();
-        assert!(text.contains("Test"));
-
-        // Verify horizontal lines are drawn
-        assert!(text.contains("─"));
-    }
-
-    #[test]
-    fn test_divider_label_clipping() {
-        // Test that label is clipped when not enough space
-        let mut buffer = Buffer::new(10, 1);
-        let area = Rect::new(0, 0, 10, 1);
-        let mut ctx = RenderContext::new(&mut buffer, area);
-
-        let d = divider().label("VeryLongLabelThatWontFit");
-        d.render(&mut ctx);
-        // Should not panic and render what fits
-    }
-
-    #[test]
-    fn test_divider_vertical_uses_vline() {
-        // Test vertical divider uses draw_vline
-        let mut buffer = Buffer::new(1, 5);
-        let area = Rect::new(0, 0, 1, 5);
-        let mut ctx = RenderContext::new(&mut buffer, area);
-
-        let d = vdivider();
-        d.render(&mut ctx);
-
-        // All cells should have vertical line
-        for y in 0..5 {
-            assert_eq!(buffer.get(0, y).map(|c| c.symbol), Some('│'));
-        }
-    }
-}
+// All tests moved to tests/widget/divider.rs
