+++ src/widget/switch.rs
@@ -446,60 +446,4 @@ pub fn toggle(label: impl Into<String>) -> Switch {
 impl_styled_view!(Switch);
 impl_props_builders!(Switch);
 
-// Most tests moved to tests/widget_tests.rs
-// Tests below access private fields and must stay inline
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_switch_label() {
-        let s = Switch::new().label("Enable feature");
-        assert_eq!(s.label, Some("Enable feature".to_string()));
-    }
-
-    #[test]
-    fn test_switch_style() {
-        let s = Switch::new().style(SwitchStyle::IOS);
-        assert!(matches!(s.style, SwitchStyle::IOS));
-    }
-
-    #[test]
-    fn test_switch_colors() {
-        let s = Switch::new().on_color(Color::CYAN).off_color(Color::RED);
-        assert_eq!(s.on_color, Color::CYAN);
-        assert_eq!(s.off_color, Color::RED);
-    }
-
-    #[test]
-    fn test_switch_custom_text() {
-        let s = Switch::new().text("Yes", "No");
-        assert_eq!(s.on_text, Some("Yes".to_string()));
-        assert_eq!(s.off_text, Some("No".to_string()));
-    }
-
-    #[test]
-    fn test_switch_label_right() {
-        let s = Switch::new().label("Test").label_right();
-        assert!(!s.label_left);
-    }
-
-    #[test]
-    fn test_toggle_helper() {
-        let s = toggle("Enable");
-        assert_eq!(s.label, Some("Enable".to_string()));
-    }
-
-    #[test]
-    fn test_switch_checked_alias() {
-        // Test checked() is an alias for on()
-        let s = Switch::new().checked(true);
-        assert!(s.is_on());
-        assert!(s.is_checked());
-
-        let s = Switch::new().checked(false);
-        assert!(!s.is_on());
-        assert!(!s.is_checked());
-    }
-}
+// All tests moved to tests/widget/switch.rs
