+++ tests/widget/divider.rs
@@ -2,6 +2,7 @@
 
 use revue::layout::Rect;
 use revue::render::Buffer;
+use revue::style::Color;
 use revue::widget::traits::RenderContext;
 use revue::widget::{divider, vdivider, View};
 
@@ -82,6 +83,42 @@ fn test_divider_vertical_uses_vline() {
     for y in 0..5 {
         assert_eq!(buffer.get(0, y).map(|c| c.symbol), Some('│'));
     }
+}
+
+// New tests from main branch - render-based verification
+#[test]
+fn test_divider_styles_render() {
+    // Test that all divider styles render without panic
+    let styles = [
+        divider(),
+        divider().dashed(),
+        divider().dotted(),
+        divider().double(),
+        divider().thick(),
+    ];
+
+    for d in styles {
+        let mut buffer = Buffer::new(20, 1);
+        let area = Rect::new(0, 0, 20, 1);
+        let mut ctx = RenderContext::new(&mut buffer, area);
+        d.render(&mut ctx);
+    }
+}
+
+#[test]
+fn test_divider_builder_chain() {
+    // Test that builder methods chain correctly
+    let mut buffer = Buffer::new(30, 1);
+    let area = Rect::new(0, 0, 30, 1);
+    let mut ctx = RenderContext::new(&mut buffer, area);
+
+    let d = divider().dashed().color(Color::RED).label("Test").margin(2);
+    d.render(&mut ctx);
+
+    let text: String = (0..30)
+        .filter_map(|x| buffer.get(x, 0).map(|c| c.symbol))
+        .collect();
+    assert!(text.contains("Test"));
 }
 
 // =============================================================================
